+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 * Removal of ``r.read/close`` methods
 * New ``r.fo`` interface for advanced response usage.*
 * Automatic expansion of parameterized headers
+* Turn off Redirects for GET/HEAD requests
 
 0.6.1 (2011-08-20)
 ++++++++++++++++++